Updated 10x tool for smarter barcode correction.
- Now the 10x tool can take an illumina bam file as an optional
argument.  This bam file will be used to seed starcode with barcodes for
correction.  The barcodes from the long reads will also be used as
evidence.  They are added with a confidence score rather than a proper
count.  This score is a function of the barcode length and base
qualities of the barcode bases.